Remove redundant code in index_sections_reader
