+++ pkg/dataobj/metastore/object_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"os"
 	"slices"
+	"sync"
 	"testing"
 	"time"
 
@@ -19,6 +20,7 @@ import (
 	"github.com/grafana/loki/v3/pkg/dataobj/consumer/logsobj"
 	"github.com/grafana/loki/v3/pkg/dataobj/index/indexobj"
 	"github.com/grafana/loki/v3/pkg/dataobj/sections/pointers"
+	"github.com/grafana/loki/v3/pkg/dataobj/sections/postings"
 	"github.com/grafana/loki/v3/pkg/dataobj/sections/streams"
 	"github.com/grafana/loki/v3/pkg/dataobj/uploader"
 	"github.com/grafana/loki/v3/pkg/logproto"
@@ -891,4 +893,97 @@ func TestIndexSectionsReader_ReaderSelection(t *testing.T) {
 			require.Equal(t, int64(1), rec.NumRows())
 		})
 	}
+}
+
+// fixtureBuildMu serializes indexobj fixture builds across parallel tests: concurrent Builder.Flush
+// calls race on the package-global defaultCompressionOptions.zstdWriter lazy init in
+// dataset/page_compress_writer.go (a pre-existing race the mutex sidesteps).
+var fixtureBuildMu sync.Mutex
+
+// newFixtureBuilder returns an indexobj.Builder configured for the reader fixtures.
+func newFixtureBuilder(t *testing.T) *indexobj.Builder {
+	t.Helper()
+
+	builder, err := indexobj.NewBuilder(logsobj.BuilderBaseConfig{
+		TargetPageSize:          1024 * 1024,
+		TargetObjectSize:        10 * 1024 * 1024,
+		TargetSectionSize:       128,
+		BufferSize:              1024 * 1024,
+		SectionStripeMergeLimit: 2,
+	}, nil)
+	require.NoError(t, err)
+	return builder
+}
+
+// flushFixture flushes builder under fixtureBuildMu (which guards the racy
+// zstdWriter lazy init reached from Flush; see fixtureBuildMu) and registers the
+// returned object's closer for cleanup.
+func flushFixture(t *testing.T, builder *indexobj.Builder) *dataobj.Object {
+	t.Helper()
+
+	fixtureBuildMu.Lock()
+	obj, closer, err := builder.Flush()
+	fixtureBuildMu.Unlock()
+	require.NoError(t, err)
+	t.Cleanup(func() { _ = closer.Close() })
+	return obj
+}
+
+func buildStreamsPointersFixture(t *testing.T) *dataobj.Object {
+	t.Helper()
+
+	builder := newFixtureBuilder(t)
+
+	_, err := builder.AppendStream(tenantID, streams.Stream{
+		ID:               1,
+		Labels:           labels.New(labels.Label{Name: "app", Value: "foo"}),
+		MinTimestamp:     now.Add(-3 * time.Hour),
+		MaxTimestamp:     now.Add(-2 * time.Hour),
+		UncompressedSize: 5,
+	})
+	require.NoError(t, err)
+	require.NoError(t, builder.ObserveLogLine(tenantID, "test-path", 0, 1, 1, now.Add(-3*time.Hour), 5))
+
+	return flushFixture(t, builder)
+}
+
+func buildCombinedFixture(t *testing.T) *dataobj.Object {
+	t.Helper()
+
+	builder := newFixtureBuilder(t)
+
+	_, err := builder.AppendStream(tenantID, streams.Stream{
+		ID:               1,
+		Labels:           labels.New(labels.Label{Name: "app", Value: "foo"}),
+		MinTimestamp:     now.Add(-3 * time.Hour),
+		MaxTimestamp:     now.Add(-2 * time.Hour),
+		UncompressedSize: 5,
+		Rows:             1,
+	})
+	require.NoError(t, err)
+
+	require.NoError(t, builder.ObserveLogLine(tenantID, "test-path", 0, 1, 1, now.Add(-3*time.Hour), 5))
+
+	builder.ObserveLabelPosting(tenantID, postings.LabelObservation{
+		ObjectPath:       "test-path",
+		SectionIndex:     0,
+		ColumnName:       "app",
+		LabelValue:       "foo",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	})
+
+	builder.PrepareBloomColumn(tenantID, "test-path", 0, "traceID", 1)
+	require.NoError(t, builder.ObserveBloomPosting(tenantID, postings.BloomObservation{
+		ObjectPath:       "test-path",
+		SectionIndex:     0,
+		ColumnName:       "traceID",
+		Value:            "abcd",
+		StreamID:         1,
+		Timestamp:        now.Add(-3 * time.Hour),
+		UncompressedSize: 5,
+	}))
+
+	return flushFixture(t, builder)
 }
